+++ src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { Routes } from "@angular/router"
 import { WelcomePage } from "./pages/welcome/welcome"
 import { ProsodyPage } from "./pages/new-poem/step-1/ProsodyPage"
 import { MivesPage } from "./pages/new-poem/step-2/MivesPage"
-import { WeightsPage } from "./pages/new-poem/step-3/weights"
+import { WeightsPage } from "./pages/new-poem/step-3/WeightsPage"
 import { PoemPage } from "./pages/poem/poem.component"
 
 export const routes: Routes = [
