+++ src/app/pages/new-poem/step-3/WeightsPage.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing"
 
-import { WeightsPage } from "./weights"
+import { WeightsPage } from "./WeightsPage"
 
 describe("Step3Component", () => {
     let component: WeightsPage
